fix: [ENH] Rename `*_v2_pkg.py` to `*_pkg_v2.py` (#2339)
