+++ docs/source/pkg_v2.rst
@@ -96,6 +96,6 @@ See the detailed API documentation for the available V2 Package classes below:
 
    models.temporal_fusion_transformer._tft_pkg_v2.TFT_pkg_v2
    models.dlinear._dlinear_pkg_v2.DLinear_pkg_v2
-   models.samformer._samformer_v2_pkg.Samformer_pkg_v2
-   models.tide._tide_dsipts._tide_v2_pkg.TIDE_pkg_v2
+   models.samformer._samformer_pkg_v2.Samformer_pkg_v2
+   models.tide._tide_dsipts._tide_pkg_v2.TIDE_pkg_v2
    models.timexer._timexer_pkg_v2.TimeXer_pkg_v2
