+++ pytorch_forecasting/models/samformer/__init__.py
@@ -4,6 +4,6 @@ DSIPTS Implementation of Samformer for V2
 """
 
 from pytorch_forecasting.models.samformer._samformer_v2 import Samformer
-from pytorch_forecasting.models.samformer._samformer_v2_pkg import Samformer_pkg_v2
+from pytorch_forecasting.models.samformer._samformer_pkg_v2 import Samformer_pkg_v2
 
 __all__ = ["Samformer", "Samformer_pkg_v2"]
