+++ pytorch_forecasting/models/samformer/_samformer_v2.py
@@ -35,7 +35,7 @@ class Samformer(BaseModel):
     @classmethod
     def _pkg(cls):
         """Return the package class for this model."""
-        from pytorch_forecasting.models.samformer._samformer_v2_pkg import (
+        from pytorch_forecasting.models.samformer._samformer_pkg_v2 import (
             Samformer_pkg_v2,
         )
