+++ pytorch_forecasting/models/tide/_tide_dsipts/__init__.py
@@ -1,6 +1,6 @@
 """DSIPTS Tide Implementation for V2"""
 
 from pytorch_forecasting.models.tide._tide_dsipts._tide_v2 import TIDE
-from pytorch_forecasting.models.tide._tide_dsipts._tide_v2_pkg import TIDE_pkg_v2
+from pytorch_forecasting.models.tide._tide_dsipts._tide_pkg_v2 import TIDE_pkg_v2
 
 __all__ = ["TIDE", "TIDE_pkg_v2"]
